change readline/promises.mjs to readline/promises.ts
